test/ai-speaker/server/requirements.txt
pydub
gTTS
python-dotenv
openai-whisper
openai
azure-identity